feat(types): centralize task dtos
