--- apps/api/src/tasks/dto/index.ts
@@ -1,4 +1,0 @@
-export * from './create-task.dto';
-export * from './query.dto';
-export * from './task-id-param.dto';
-export * from './update-task.dto';
