--- apps/api/src/tasks/dto/task-id-param.dto.ts
@@ -1,6 +1,0 @@
-import { IsUUID } from 'class-validator';
-
-export class TaskIdParamDto {
-  @IsUUID()
-  id: string;
-}
