--- apps/api/src/tasks/dto/update-task.dto.ts
@@ -1,42 +1,0 @@
-import { Type } from 'class-transformer';
-import {
-  IsArray,
-  IsDateString,
-  IsEnum,
-  IsNotEmpty,
-  IsOptional,
-  IsString,
-  ValidateNested,
-} from 'class-validator';
-import type { UpdateTask } from '@repo/types';
-import { TaskPriority, TaskStatus } from '@repo/types';
-import { TaskAssigneeDto } from './create-task.dto';
-
-export class UpdateTaskDto implements UpdateTask {
-  @IsOptional()
-  @IsString()
-  @IsNotEmpty()
-  title?: string;
-
-  @IsOptional()
-  @IsString()
-  description?: string | null;
-
-  @IsOptional()
-  @IsEnum(TaskStatus)
-  status?: TaskStatus;
-
-  @IsOptional()
-  @IsEnum(TaskPriority)
-  priority?: TaskPriority;
-
-  @IsOptional()
-  @IsDateString()
-  dueDate?: string | null;
-
-  @IsOptional()
-  @IsArray()
-  @ValidateNested({ each: true })
-  @Type(() => TaskAssigneeDto)
-  assignees?: TaskAssigneeDto[];
-}
