+++ packages/types/package.json
@@ -25,6 +25,10 @@
     "check-types": "tsc --noEmit",
     "prepare": "pnpm run build"
   },
+  "dependencies": {
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.14.2"
+  },
   "devDependencies": {
     "@repo/eslint-config": "workspace:*",
     "@repo/typescript-config": "workspace:*",
